[CCAP-974] - removed duplicate tests and combined into 3 main types
