+++ src/main/java/org/ilgcc/app/pdf/ApplicationPreparer.java
@@ -65,9 +65,6 @@ public class ApplicationPreparer implements SubmissionFieldPreparer {
                 + PreparerUtilities.numberValueOf(unemploymentIncome) + PreparerUtilities.numberValueOf(royaltiesIncome)
                 + PreparerUtilities.numberValueOf(pensionIncome) + PreparerUtilities.numberValueOf(workersIncome);
 
-        results.put("childcareStartDate",
-                new SingleField("childcareStartDate", inputData.get("earliestChildcareStartDate").toString(), null));
-
         results.put("otherMonthlyIncomeApplicant",
                 new SingleField("otherMonthlyIncomeApplicant", String.format("%.0f", Math.floor(totalExpenses)), null));
 
@@ -101,6 +98,9 @@ public class ApplicationPreparer implements SubmissionFieldPreparer {
             results.put("referralServicesPhysicalOrMentalDisability",
                     new SingleField("referralServicesPhysicalOrMentalDisability", "true", null));
         }
+
+        results.put("clientResponseConfirmationCode",
+                new SingleField("clientResponseConfirmationCode", submission.getShortCode(), null));
 
         return results;
     }
